refactor: migrating returns correct amount, 4.8
